@@ -7,23 +7,23 @@
   },
   "version": "0.1.0",
   "description": "React Authentication with Redux Thunk",
-  "homepage": "https://github.com/feesler/react-redux-auth",
+  "homepage": "https://henryfeesler.com/netology/11/react-redux-auth",
   "private": true,
   "dependencies": {
-    "@reduxjs/toolkit": "^1.5.0",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.8.3",
-    "classnames": "^2.2.6",
+    "@reduxjs/toolkit": "^1.9.7",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
+    "@testing-library/user-event": "^14.5.2",
+    "classnames": "^2.5.1",
     "lodash": "^4.17.21",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.3",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.0.5",
-    "web-vitals": "^1.1.1"
+    "prop-types": "^15.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^9.1.0",
+    "react-router-dom": "^6.22.1",
+    "react-scripts": "^5.0.1",
+    "redux": "^5.0.1",
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "prestart": "npm install",
