@@ -1,5 +1,5 @@
 import React from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import Banner from './Banner.jsx';
 import NewsList from './NewsList.jsx';
 import Page404 from './Page404.jsx';
@@ -7,11 +7,11 @@ import Page404 from './Page404.jsx';
 function Content() {
   return (
     <div className="content">
-      <Switch>
+      <Routes>
         <Route path="/" exact component={Banner} />
         <Route path="/news" exact component={NewsList} />
         <Route component={Page404} />
-      </Switch>
+      </Routes>
     </div>
   )
 }
