@@ -1,5 +1,5 @@
 import React from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 import LoginForm from './LoginForm.jsx';
 import UserProfile from './UserProfile.jsx';
 
@@ -7,10 +7,10 @@ function Header() {
   return (
     <div className="header">
       <div className="header__logo me-4">Neto Social</div>
-      <Switch>
+      <Routes>
         <Route path="/" exact component={LoginForm} />
         <Route component={UserProfile} />
-      </Switch>
+      </Routes>
     </div>
   )
 }
